Fixed app crashes due to NULL entries in vtables of implementations of pure virtual classes
When implementing a jaenerated pure virtual class , our application would crash - but not always.

The cause: Class.getDeclaredMethods() isn't sorted (read the docs!),and the BridJ code that lists virtual methods before building the vtable only worked properly when the bridj.Pointer overload the the virtual method appears after the native method.

The fix prefers to add the the bridj.Pointer-based overload t the list of virtuals, resulting in a complete vtable later on.

Of course one could just delete the native entries in the jaenerated code, but when the bridjed framework has far too many classes, and needs frequent updates, this just gets tedious.